requirements.txt
pandas==2.2.3
mysql-connector-python==9.2.0
SQLAlchemy==2.0.23
openpyxl==3.1.5